chore: in mem store implements revoke reason too
